requirements.txt
# KSU IT RAG Chatbot - Python Dependencies

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
trafilatura>=1.7.0

# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML/AI
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.2.0
numpy>=1.24.3

# Vector Database
# Pinecone (cloud-based, for team collaboration)
# Requires PINECONE_API_KEY and PINECONE_ENVIRONMENT in .env file
pinecone-client==2.2.4

# Note: ChromaDB kept for local development/testing if needed
# chromadb==0.4.18

# Data Processing
pandas==2.1.3
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0

# Optional: OpenAI API (if using OpenAI instead of Hugging Face)
# openai==1.3.5

